Restructure gen-c transformer into three parts.
